+++ files/assets/tvs/multitv/HISTORY.md
@@ -1,6 +1,15 @@
 History
 ================================================================================
-- 1.5.3 May 15, 2013
+
+- 1.5.5 June 8, 2013
+    - Snippet: `reverse` parameter
+    
+- 1.5.4 May 24, 2013
+    - Snippet: bugfix for empty multiTV
+    - CustomTV: bugfix for datatables and template var id
+    - CustomTV: corrected invalid language settings
+
+- 1.5.3 May 24, 2013
     - CustomTV: bugfix for row reordering problem in datatables with more than one datatables multiTV
 
 - 1.5.2 May 15, 2013